first pass of docker + ci builds for ubuntu
