+++ arm64.yml
@@ -1,0 +1,9 @@
+version: '3.2'
+
+services:
+   imagebuild:
+     build: 
+       args:
+         IMAGE_ARCH: arm64v8
+         QEMU_FILE: qemu-aarch64-static
+         RUN_CMD: /input/ubuntu.sh arm64
