--- assets/all/assets.txt
@@ -1,9 +1,0 @@
-addNonRootUser.sh 1549996495 734f0d0a122c8b283f7dbad94a94f49c
-extractFilesystem.sh 1549996495 8997eb9d73ba810becd2dd6c8f38b037
-ld.so.preload 1532022179 43618bdfa97456206c3f21bd8fc2a9cb
-nosudo 1549996495 6c77f57a4be314c57682240754d85327
-startSSHServer.sh 1549996495 1455ac210cfa0966e3921d68bd4957d7
-startVNCServer.sh 1549996495 a52afb25bd9b9ef7ea87c6de3d88b3ad
-startVNCServerStep2.sh 1549996495 c30b0a63f1073c601e1fef1fc4503c8e
-startXSDLServer.sh 1549996495 e044b2cddb0074da74b983fc94ca0c8c
-userland_profile.sh 1549996495 8c185e7210b2af78d1f6135636737cde
