--- assets/arm/assets.txt
@@ -1,6 +1,0 @@
-busybox 1539642474 100662e8f98bdc6d527418c1487387d4
-libdisableselinux.so 1549996495 f6b71761886e023c9a89850bd16208ff
-rootfs.tar.gz.part00 1549996495 6320806e5645b8a8eedce2b556c5dcf4
-rootfs.tar.gz.part01 1549996496 192a86d20b2e4a91cf6e26b8f620027e
-rootfs.tar.gz.part02 1532022247 d41d8cd98f00b204e9800998ecf8427e
-rootfs.tar.gz.part03 1532022249 d41d8cd98f00b204e9800998ecf8427e
