--- assets/arm64/assets.txt
@@ -1,6 +1,0 @@
-busybox 1539642775 644b9db2e7e908e25f0b78ad1fd3639e
-libdisableselinux.so 1549996496 231ea1d25f3ce2ed8c5db3dbfc703588
-rootfs.tar.gz.part00 1549996496 d044a6ea54ac084899117f3d29da1aa7
-rootfs.tar.gz.part01 1549996496 b5fcde234e94119dbaf5f75aa3ca277d
-rootfs.tar.gz.part02 1532022292 d41d8cd98f00b204e9800998ecf8427e
-rootfs.tar.gz.part03 1532022293 d41d8cd98f00b204e9800998ecf8427e
