--- assets/x86/assets.txt
@@ -1,6 +1,0 @@
-busybox 1539642672 89396ad50debe6470893e85f411f7936
-libdisableselinux.so 1549996496 5b32ab2e28269559a8598eadfaec4cf9
-rootfs.tar.gz.part00 1549996497 442871ec60f19594dc94966532afcdaf
-rootfs.tar.gz.part01 1549996497 c86be081742a44071ca4c8e902659f3f
-rootfs.tar.gz.part02 1549996497 e1c9be7c3605d8b7d93ce8eb61148926
-rootfs.tar.gz.part03 1532022326 d41d8cd98f00b204e9800998ecf8427e
