--- assets/x86_64/assets.txt
@@ -1,6 +1,0 @@
-busybox 1539642700 df0b535de05540ee9de05d035910351d
-libdisableselinux.so 1549996497 a64be1791187c1edcd53b6a71e3ae362
-rootfs.tar.gz.part00 1549996498 718acabceedb749318c95d8f8deffc9f
-rootfs.tar.gz.part01 1549996498 f46c981574753050f911942986f9ab34
-rootfs.tar.gz.part02 1549996498 e119d0e3e90dbddd3c3d34be4cd2ea7a
-rootfs.tar.gz.part03 1532022358 d41d8cd98f00b204e9800998ecf8427e
