+++ x86.yml
@@ -1,0 +1,9 @@
+version: '3.2'
+
+services:
+   imagebuild:
+     build: 
+       args:
+         IMAGE_ARCH: i386
+         QEMU_FILE: qemu-i386-static
+         RUN_CMD: /input/ubuntu.sh x86
